Change to how strlcpy and strlcat are handled
Put strlcpy and strlcat object files back in the source group in lib.
Moved libfko_util.a to the common directory (though sources remain in
lib).  Client and server code looks to common dir for libfko-util.
This fixes issue with strlcpy showing as undefined symbol when perl FKO
module is loaded.