[test suite] bug fix for rotate digest cache tests
When the test suite is executed with '--include "rotate"' then previous tests
aren't executed in order to create a new digest cache file.  So, when init() is
called and a clean slate is established, there is nothing to rotate away.  This
change creates the default digest cache data (comment line only) if the file
doesn't already exist for the rotate tests.