[libfko] change SPA packet format to use true 16 bytes of random data
This is a significant change that updates the current strategy of 16
digits of random data (which works out to about 53 bits of entropy) to
using true 16 bytes of random data.  By itself, this change breaks
backwards compatibility with prior versions of fwknop.  However, a
subsequent change will add a legacy mode (similar to the encryption
legacy mode) that retains backwards compatibility for those that need
it.