[test suite] Added valgrind suppressions for gpgme
Running the test suite with --enable-valgrind resulted in large numbers
of leaks detected in gpgme functions.  This commit adds a valgrind
suppressions file to squash these errors (which are not fwknop's fault),
and also enables the valgrind --child-slient-after-fork option by
default.  Both of these can disable in test suite execution with two
new options: --valgrind-disable-suppressions and
--valgrind-disable-child-silent.